requirements.txt
crewai
crewai_tools
openai
python-dotenv
langchain-community
langchain
langchain-openai
ipython
serpapi
google-search-results